+++ src/main/java/org/example/androidbackend/services/GenreService.java
@@ -2,5 +2,5 @@ package org.example.androidbackend.services;
 
 public interface GenreService {
     boolean addGenre(String genreName);
-    boolean updateGenre(int id, String genreName);
+    boolean updateGenre(Long id, String genreName);
 }
